Support printing default TOML config
Allow Linux distributions that want to package the default configuration
to be able to generate the configuration statically for their packages
rather than requiring it to be generated at first start time. This also
allows for distro config management systems to provide better support
for changes to the default over time.